Refactor track-changer slightly
