+++ app/css/components/mentor-track-selector.css
@@ -1,0 +1,94 @@
+@import "../styles";
+
+.c-mentor-track-selector {
+    & .c-search-bar {
+        @apply mb-32;
+
+        & .--search {
+            width: 480px;
+            margin-right: auto;
+        }
+
+        & .selected {
+            @apply text-16 font-medium;
+            @apply flex-shrink-0;
+
+            &.none {
+                @apply text-textColor6;
+            }
+        }
+    }
+
+    & .tracks {
+        @apply mb-40;
+        @apply grid gap-20;
+        grid-template-columns: repeat(4, auto);
+
+        & .track {
+            & label {
+                @apply bg-backgroundColorB shadow-lg rounded-8;
+                @apply py-24;
+                @apply flex flex-col items-center;
+                @apply relative;
+                @apply border-2 border-transparent;
+
+                cursor: pointer;
+                transition: background 0.2s ease-out, box-shadow 0.2s ease-out;
+
+                &:hover {
+                    @apply shadow-lgZ1;
+                    @apply bg-backgroundColorA;
+                }
+
+                & .selected-icon {
+                    position: absolute;
+                    top: -20px;
+                    right: -20px;
+
+                    @apply grid place-items-center;
+                    @apply bg-darkBlue rounded-circle;
+                    width: 40px;
+                    height: 40px;
+
+                    @apply hidden;
+
+                    & .c-icon {
+                        height: 20px;
+                        width: 20px;
+                    }
+                }
+
+                & .c-track-icon {
+                    height: 64px;
+                    width: 64px;
+                    @apply mb-12;
+                }
+
+                & .title {
+                    @apply text-18 leading-140 text-textColor3 font-semibold;
+                    @apply mb-6;
+                }
+
+                & .info {
+                    @apply text-14 leading-160 text-textColor6;
+                }
+            }
+
+            & input[type="checkbox"] {
+                opacity: 0;
+                &:checked + label {
+                    @apply border-darkBlue;
+                    @apply bg-veryLightBlue;
+
+                    & .selected-icon {
+                        @apply grid;
+                    }
+                }
+
+                &.focus-visible + label {
+                    @apply border-2 border-inputBorderColorFocus;
+                }
+            }
+        }
+    }
+}
