+++ app/javascript/components/mentoring/TrackSelector.tsx
@@ -1,0 +1,71 @@
+import React, { useCallback } from 'react'
+import { usePaginatedRequestQuery } from '../../hooks/request-query'
+import { useIsMounted } from 'use-is-mounted'
+import { TracksList } from './track-selector/TracksList'
+import { SelectedTracksMessage } from './track-selector/SelectedTracksMessage'
+import { ContinueButton } from './track-selector/ContinueButton'
+import { SearchBar } from './track-selector/SearchBar'
+import { useList } from '../../hooks/use-list'
+
+export type APIResponse = {
+  tracks: readonly Track[]
+}
+
+export type Track = {
+  id: string
+  title: string
+  iconUrl: string
+  avgWaitTime: string
+  numSolutionsQueued: number
+}
+
+export const TrackSelector = ({
+  tracksEndpoint,
+  selected,
+  setSelected,
+  onContinue,
+}: {
+  tracksEndpoint: string
+  selected: string[]
+  setSelected: (selected: string[]) => void
+  onContinue: () => void
+}): JSX.Element => {
+  const isMountedRef = useIsMounted()
+  const { request, setCriteria } = useList({
+    endpoint: tracksEndpoint,
+    options: {},
+  })
+  const { status, resolvedData, isFetching, error } = usePaginatedRequestQuery<
+    APIResponse
+  >('tracks', request, isMountedRef)
+
+  const handleContinue = useCallback(() => {
+    onContinue()
+  }, [onContinue])
+
+  return (
+    <div className="c-mentor-track-selector">
+      <div className="c-search-bar">
+        <SearchBar
+          value={request.query.criteria || ''}
+          setValue={setCriteria}
+        />
+        {isFetching ? <span>Fetching</span> : null}
+        <SelectedTracksMessage numSelected={selected.length} />
+        <ContinueButton
+          disabled={selected.length === 0}
+          onClick={handleContinue}
+        />
+      </div>
+      <div className="tracks">
+        <TracksList
+          status={status}
+          selected={selected}
+          setSelected={setSelected}
+          data={resolvedData}
+          error={error}
+        />
+      </div>
+    </div>
+  )
+}
