+++ app/javascript/packs/application.tsx
@@ -62,6 +62,7 @@ import '../../css/components/iteration-pane'
 import '../../css/components/explainer'
 import '../../css/components/markdown-editor'
 import '../../css/components/mentor-discussion-summary'
+import '../../css/components/mentor-track-selector'
 import '../../css/components/tag'
 
 import '../../css/components/widgets/exercise'
@@ -72,6 +73,7 @@ import '../../css/modals/mentoring-sessions'
 import '../../css/modals/finish-mentor-discussion'
 import '../../css/modals/welcome-to-v3'
 import '../../css/modals/become-mentor'
+import '../../css/modals/change-mentor-tracks'
 
 import '../../css/dropdowns/notifications'
 import '../../css/dropdowns/reputation'
